test: add specs (#55)
